requirements.txt
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PlexAPI==4.11.2
pydantic==1.9.1
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.37
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
